// src/types/property.ts
export interface Property {
    id: number;
    images: Array<string>;
    price: string;
    title: string;
    location: string;
    bedrooms: number;
    bathrooms: number;
    lot_area: string;
    offer_type: string;
    city_town: string;
    province: string;
    property_type: string;
    description: string;
    floor_area: string;
    type: string;
}